+++ packages/platform-app/src/platformApps.spec.ts
@@ -1,0 +1,7 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+import { describe, it } from 'vitest';
+
+describe('Platform App', () => {
+    it('Define App', async () => {});
+});