+++ packages/app-bridge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @frontify/app-bridge
 
+## 4.0.0-alpha.1
+
+### Patch Changes
+
+- [#742](https://github.com/Frontify/brand-sdk/pull/742) [`5ea66a7`](https://github.com/Frontify/brand-sdk/commit/5ea66a79bac131f5b34f1a1bf60fbe81add6b17f) Thanks [@anxobotana](https://github.com/anxobotana)! - feat: add context.template to AppBridgeTheme
+
+## 3.2.0
+
+### Minor Changes
+
+- [#731](https://github.com/Frontify/brand-sdk/pull/731) [`b9c4fc4`](https://github.com/Frontify/brand-sdk/commit/b9c4fc4fddacabb988ceb0cbe125c1793a697e19) Thanks [@anxobotana](https://github.com/anxobotana)! - feat: add theme template context to appBridgeThemes context
+
 ## 4.0.0-alpha.0
 
 ### Major Changes